Auto set Render Resolution to match Intermediate/Proxy resolution
Found a hack which solves this, however I can't get auto "Fit Preview in Window". to work - would be nice though. This way doesn't seem to work anymore: https://blender.stackexchange.com/questions/118741/how-to-make-the-sequencer-preview-area-fullscreen-with-python

Maybe it would be more user friendly to add the proxy files to a PROXY folder and same thing with the intermediates, so the source folder will stay clean?